% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_trials.R
\name{combine_trials}
\alias{combine_trials}
\title{The constructor for the \code{combined_trials} S3 class}
\usage{
combine_trials(
  current_trial,
  external_trial,
  response,
  covariates,
  group,
  group_levels = c("Treatment", "Control", "External")
)
}
\arguments{
\item{current_trial}{The current trial data frame.}

\item{external_trial}{The external trial data frame.}

\item{response}{Tidy-select expression for the response variable.}

\item{covariates}{Tidy-select expression for the covariates.}

\item{group}{Tidy-select expression for the group variable.}
}
\value{
An object of class \code{combined_trials}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
tb <- combine_trials(dat1$trial_data, dat1$external_data,
    response = Y,
    covariates = c(x1, x2, x3),
    group = group)
}
