% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat1}
\alias{dat1}
\alias{dat2}
\title{Example data sets}
\format{
\subsection{\code{dat1}}{

A list of two data frames:
\describe{
\item{x1, x2, x3}{Covarites}
\item{Y}{Fesponse}
\item{group}{Treatment group}
}
}

\subsection{\code{dat2}}{

A list of two data frames:
\describe{
\item{x1, x2}{Covarites}
\item{Y}{Fesponse}
\item{group}{Treatment group}
}
}
}
\usage{
dat1
}
\description{
\code{dat1} reflects the Unmatched case while \code{dat2} reflects
the Unmatched-Unmeasured case.
}
\keyword{datasets}
