% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ps.R
\name{estimate_ps}
\alias{estimate_ps}
\title{Estimate propensity scores}
\usage{
estimate_ps(object, formula, method = "glm", ...)
}
\arguments{
\item{object}{A combined_trials object.}

\item{formula}{A formula specifying the PS model.}

\item{method}{The estimation method ("glm", "cbps", "gbm", "superlearner").}

\item{...}{Additional arguments passed to the underlying estimation function.}
}
\value{
The updated object with propensity scores stored in \code{ps}.
}
\description{
Estimate propensity scores
}
