% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maPP.R
\name{maPP}
\alias{maPP}
\title{Title}
\usage{
maPP(object, A, times, ...)
}
\arguments{
\item{object}{a combined_trials object}

\item{A}{the nominal number of patients to be borrowed}

\item{times}{the number of samples used to find the template}

\item{...}{}
}
\value{
The result of class maPP_result
}
\description{
Title
}
