% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSborrowing-package.R
\docType{package}
\name{PSborrowing-package}
\alias{PSborrowing}
\alias{PSborrowing-package}
\title{PSborrowing: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MinLinSTAT/PSborrowing}
  \item Report bugs at \url{https://github.com/MinLinSTAT/PSborrowing/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
