% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssPP.R
\name{pssPP}
\alias{pssPP}
\title{Propensity-score stratified power prior (Wang et al. 2019)
Stratify using the whole current RCT following Chen et al. (2020) and
use equal weights to combine}
\usage{
pssPP(object, A, ...)
}
\arguments{
\item{object}{An object of class ps_stratified}

\item{A}{The nominal number of patients to be borrowed}
}
\description{
Propensity-score stratified power prior (Wang et al. 2019)
Stratify using the whole current RCT following Chen et al. (2020) and
use equal weights to combine
}
