% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_divide.R
\name{summary.ps_stratified}
\alias{summary.ps_stratified}
\title{Compute the Stratum-Specific Summary Statistics}
\usage{
\method{summary}{ps_stratified}(object)
}
\arguments{
\item{combined_trials}{An object of class \code{combined_trials}.}
}
\value{
A data frame with stratum-specific statistics.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
