% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_script.R
\name{use_script}
\alias{use_script}
\title{Create or Open a Script in the \code{inst/scripts} Directory}
\usage{
use_script(filename)
}
\arguments{
\item{filename}{A character string specifying the name of the script file (without
the \code{.R} extension) to create or open within \verb{inst/scripts/}.}
}
\value{
This function is called for its side effects and returns \code{NULL} invisibly.
It creates or opens a file in the \verb{inst/scripts/} directory, and if running in
RStudio, opens it in the editor.
}
\description{
This function checks if the \verb{/inst/scripts/} directory exists in the package.
If it doesn't, the function creates the directory. Then, it checks if a script
file with the specified name already exists in that directory. If the file exists,
it opens the file in RStudio. Otherwise, it creates a new file and opens it.
}
\details{
This function is intended to streamline script creation within a package structure,
making it easy to create or edit scripts in the \verb{inst/scripts/} directory.
If \code{rstudioapi} is available and the function is called interactively in RStudio,
the script file will open automatically in the editor.
}
\examples{
\dontrun{
# Create or open a script named "my_script.R" in the inst/scripts directory
use_script("my_script")
}

}
